move FTS to its own folder/namespace & cleanup form code
