+++ poshsecframework/FirstTimeSetup/Steps.cs
@@ -1,0 +1,11 @@
+namespace PoshSec.Framework.FirstTimeSetup
+{
+    public enum Steps
+    { 
+        Check_Settings = 0,
+        InitialDownload,
+        Unblock_Files,
+        Update_Help,
+        Set_Execution_Policy
+    }
+}
